requirements.txt
fastapi
uvicorn
pinecone-client
requests
sentence-transformers
python-multipart